feat: finish moving definitions and start exprs
