+++ crates/aiken-lang/src/parser/definitions/function.rs
@@ -1,8 +1,9 @@
 use chumsky::prelude::*;
 
 use crate::{
-    ast, expr,
-    parser::{error::ParseError, token::Token, utils},
+    ast,
+    expr::UntypedExpr,
+    parser::{annotation, error::ParseError, expr, token::Token, utils},
 };
 
 pub fn parser() -> impl Parser<Token, ast::UntypedDefinition, Error = ParseError> {
@@ -11,15 +12,15 @@ pub fn parser() -> impl Parser<Token, ast::UntypedDefinition, Error = ParseError
         .then_ignore(just(Token::Fn))
         .then(select! {Token::Name {name} => name})
         .then(
-            fn_param_parser(false)
+            param(false)
                 .separated_by(just(Token::Comma))
                 .allow_trailing()
                 .delimited_by(just(Token::LeftParen), just(Token::RightParen))
                 .map_with_span(|arguments, span| (arguments, span)),
         )
-        .then(just(Token::RArrow).ignore_then(type_parser()).or_not())
+        .then(just(Token::RArrow).ignore_then(annotation()).or_not())
         .then(
-            expr_seq_parser()
+            expr::sequence()
                 .or_not()
                 .delimited_by(just(Token::LeftBrace), just(Token::RightBrace)),
         )
@@ -27,7 +28,7 @@ pub fn parser() -> impl Parser<Token, ast::UntypedDefinition, Error = ParseError
             |((((opt_pub, name), (arguments, args_span)), return_annotation), body), span| {
                 ast::UntypedDefinition::Fn(ast::Function {
                     arguments,
-                    body: body.unwrap_or_else(|| expr::UntypedExpr::todo(span, None)),
+                    body: body.unwrap_or_else(|| UntypedExpr::todo(span, None)),
                     doc: None,
                     location: ast::Span {
                         start: span.start,
@@ -45,4 +46,44 @@ pub fn parser() -> impl Parser<Token, ast::UntypedDefinition, Error = ParseError
                 })
             },
         )
+}
+
+pub fn param(is_validator_param: bool) -> impl Parser<Token, ast::UntypedArg, Error = ParseError> {
+    choice((
+        select! {Token::Name {name} => name}
+            .then(select! {Token::DiscardName {name} => name})
+            .map_with_span(|(label, name), span| ast::ArgName::Discarded {
+                label,
+                name,
+                location: span,
+            }),
+        select! {Token::DiscardName {name} => name}.map_with_span(|name, span| {
+            ast::ArgName::Discarded {
+                label: name.clone(),
+                name,
+                location: span,
+            }
+        }),
+        select! {Token::Name {name} => name}
+            .then(select! {Token::Name {name} => name})
+            .map_with_span(move |(label, name), span| ast::ArgName::Named {
+                label,
+                name,
+                location: span,
+                is_validator_param,
+            }),
+        select! {Token::Name {name} => name}.map_with_span(move |name, span| ast::ArgName::Named {
+            label: name.clone(),
+            name,
+            location: span,
+            is_validator_param,
+        }),
+    ))
+    .then(just(Token::Colon).ignore_then(annotation()).or_not())
+    .map_with_span(|(arg_name, annotation), span| ast::Arg {
+        location: span,
+        annotation,
+        tipo: (),
+        arg_name,
+    })
 }
