+++ crates/aiken-lang/src/parser/expr/mod.rs
@@ -1,0 +1,3 @@
+mod sequence;
+
+pub use sequence::parser as sequence;
